+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+### Added
+* Add holidays from Keppres Nomor 10 Tahun 2024.
+* Add holidays from Keppres Nomor 33 Tahun 2024.
+
 ## 0.0.2
 ### Added
 * Add 2023 holiday data.